Section 2 Video 4
I was working with an other (a private) project, so I forgot what I did with this, which that is not very good. But good news is, I have found myself a progressive way to tackle programming in general to learn and not forget stuff in the near future. And the simple answer is just by reading the docs of the programming language instead of random websites how a for loop or syntax is constructed... Most documents has in-depth explanations what x does.

See these useful documentation links:
 - https://www.w3.org/standards/
- https://developer.mozilla.org/en-US/docs/Web/HTML
- https://developer.mozilla.org/en-US/docs/Web/CSS
- https://devdocs.io/css/

Found those links from: https://www.reddit.com/r/webdev/comments/197b8du/html_and_css_documentation/